Add table of contents workaround for h1/h2 indenting
Following discussion here:
- https://gohugo.io/content-management/toc/
- https://github.com/gohugoio/hugo/issues/1778

It is clear to me that the Hugo team has no workable solution
for how table of contents generation plays with (or rather doesn't)
markdown headers. You're not supposed to have more than 1 <h1> per
page but the ToC generator creates a top level item for <h1> elements
that don't have to exist. This is kind of frustrating because if
following the spec and doing "what you're supposed to do," your
table of contents will end up looking (subtly) wrong.

This is the suggestion from @HenrySkup on GitHub, as listed here:
https://github.com/gohugoio/hugo/issues/1778#issuecomment-522222658